Bin_history/balancing/AenerShark_Website/requirements.txt
fastapi
uvicorn[standard]
paho-mqtt
pyserial
RPi.GPIO
